requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
docx2txt>=0.8
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
langchain==0.1.20
langchain-openai==0.0.8
langchain-community==0.0.38
langsmith==0.1.147
openai>=1.0.0
# chromadb>=0.4.0  # Temporarily disabled due to compatibility issues
# faiss-cpu>=1.8.0  # Temporarily disabled due to compatibility issues
flask>=2.3.0
flask-cors>=4.0.0
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
